# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API Documentation",
        "description": "API Documentation for Product Management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/category-products": {
            "get": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Get all category products",
                "operationId": "6e01d4b73bf5dc2669f3c6747a6f3872",
                "responses": {
                    "200": {
                        "description": "List of category products"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Create a new category product",
                "operationId": "f4d44096ce183d99a948aaffd1e4234d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Electronics"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category product created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/category-products/{id}": {
            "get": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Get category product by ID",
                "operationId": "746bdc2a1fd06a210fa4542e763a249a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category product details"
                    },
                    "404": {
                        "description": "Category product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Update category product by ID",
                "operationId": "9ea8e9fc032137450870e76c151f6cc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "New Category Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category product updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Category product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category Products"
                ],
                "summary": "Delete category product by ID",
                "operationId": "733731290feb198d569c9f274a60c245",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "application/json"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "application/json"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}